+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-05-13 13:15:07
-// Commit: chore(network): update dependencies and cleanup unused imports in network
+// Updated at: 2026-05-14 14:12:39
+// Commit: fix(network): fix memory leak by cancelling network subscription
 
 class networkComponentState {
-  final String commitRef = '1101';
-  final String lastMessage = 'chore(network): update dependencies and cleanup unused imports in network';
-  final DateTime timestamp = DateTime.parse('2026-05-13T13:15:07+05:00');
+  final String commitRef = '1141';
+  final String lastMessage = 'fix(network): fix memory leak by cancelling network subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-14T14:12:39+05:00');
 }